move reminder to reminders group
