+++ src/commands/info/reminders.ts
@@ -1,0 +1,9 @@
+import { CommandInfo } from '../../classes/CommandInfo';
+
+export const info: CommandInfo = {
+   name: 'Reminders',
+   description: 'set and view reminders',
+   aliases: [],
+   commands: [],
+   paginate: true
+};
